requirements.txt
streamlit
sentence-transformers
pinecone-client
langchain
langchain-community
langchain-pinecone
openai
gitpython